pyproject.toml
[project]
name = "bgg-search-telegram-bot"
version = "0.1.0"
description = "Telegram bot to search for games on BoardGameGeek and quickly link to them"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "João Caetano", email = "joaomartinscaetano@proton.me"},
]
dependencies = [
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
upload_to_release = true
build_command = "uv build"

[tool.semantic_release.changelog]
exclude_commit_patterns = []

[tool.semantic_release.changelog.default_templates]
output_format = "md"
changelog_file = "CHANGELOG.md"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "refactor", "style", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=. --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "types-requests>=2.32.4.20250913",
]
